Add builtin documentation for remaining graphics funcs
Add builtin documentation for remaining graphic functions (part 2 of 2).
These are `dash`, `linecap`, `text` and `font`.

Change JS `linecap` to default to `"round"` rather than `"butt"`.
Round makes for nicer polylines.

Update text sample on playground. Not reading Harry Potter anymore so
the cheering charm feels stale.

Ensure that the `dash` pattern is reset and restored when drawing grid -
prior to this change the grid would be dashed too. oups.

Here are graphics in the context of builtin docs:

✅ I/O
✅ Types
✅ Map
✅ Program Control
✅ Conversion
✅ Error
✅ String
✅ Random
✅ Math
👉 Graphics
⬜️ Event Handlers