🗺️ eval: Type empty array and map lit on assignment (#171)
Type empty array and map lit on assignment with type of LHS asignable
and as []any or {}any for inferred declaration.

	a := []
	print (typeof a) // before: [] now: []any
	b:[]num
	print (typeof b) // before: []num now: []num
	b = []
	print (typeof b) // before: [] now: []num
	print (typeof []) // before: [] now: []

This seems to be more accurate and expected behaviour.

Fix the same behaviour for parameter passing and including for variadic
parameters.

Throw in a little bug fix for variadic args for good measure.

Pull-Request: https://github.com/foxygoat/evy/pull/171