evaluator: Rename runtime error to panic
Rename runtime error to panic so that index out of bounds and similar
runtime errors appear like a panic. Then all unrecoverable runtime
errors are panics or internal errors (which should never occur). While
at it, fix the error message for missing closing quote of string
literals and improve on the error section in builtins.md. This is
precursory work to finish off the last part of the sepc revision,
including errors.