docs: Add expect_err flag to doctest
Add an `expect_err` flag to go after the `evy` language marker used by
doctest to denote that the evy program is expected to exit with an
error. In these cases, we do not report the error and instead capture
the error output to put in the `evy:output` block.

Update `docs/builtins.md` to add this `expect_err` flag to the `exit`
builtin which exits with an error. This cleans up the output of
`make doctest` so it no longer reports an expected error.