✅ assert: Rework assert utility for tinygo 0.28.1 (#189)
Rework testing utility package `assert` to make use of the tinygo's new
reflection. This means we can get rid of of most code in this
packages. As expected, the size of frontend/evy.wasm stays the same.

While at it, add a couple of godoc comments to `assert.NoError` and
`assert.Equal`.

Pull-Request: https://github.com/foxygoat/evy/pull/189