lexer: Remove lexer.Run function
Remove lexer.Run function as it is only used in tests and in a hidden
command in the evy CLI. The name doesn't seem fitting either, better
would be DumpString() or so, but it actually also seems to make more
sense to stream on CLI rather than generate a full string first. This
Run function was a leftover from the early days where we debugged the
lexer with tinygo / wasm.