parser: Rename Bool AST node to BoolLiteral
Rename Bool AST node to BoolLiteral to be consistent with NumLiteral,
StringLiteral, ArrayLiteral etc. Initially Bool was named Bool and not
BoolLiteral on purpose because there are only two bool literals: true
and false which are stored explicitly rather than keeping track of the
literal value. However, this distinction causes more confusion than it
helps so keep things consistent.