evaluator: Remove DefaultBuiltins
Remove DefaultBuiltins as the idea of byo builtins is not well thought
out yet. There is only one set of builtins, if anybody wants to byo
builtins they need to fork atm. This idea came about when revisiting
the docs for the evaluator and realising that the public interface
seems way to large to support into the future.

Along with the DefaulBuiltin removal buitins have become all private, in
a follow up commit we will make the evalutor.Value interface an its
implementations private too.

This is all work towards a public API that we're willing to support and
be backwards compatible with.