evaluator: Evaluate with late composite literal typing
Evaluate with late composite literal typing, this should finally allow
for the implementation of a function like `font props:{}any` that can
also be called with a `{}num`, e.g.: `font {size:2}`. This is a really
overwhelming commit and maybe it's time to take a step back and rework
this completely, but what we have here is better than what was there
before I believe.

Unfortunately there is still a weirdness around:

	a := [[]]
	print (typeof a[0])  // should []any but is []

 I left the TODOs for the corresponding tests in the evaluator.