all: Improve ErrVarNotSet error
Improve ErrVarNotSet error by changing it from an internal error to a
runtime panic as there are legitimate ways in which to parse code that
then runs using an unset variable. The new testcase covers this. While
at it improve the error message and location where the error is
reported.