samples: CameCase all funcs and vars
CameCase all function and variable names. The decision has been made for evy
to be an on-ramp for Go and possibly also to use Go's convention of captial
letters for exported functions, to be seen. It was a hard decision because
snake case is more readable, however, underscores are harder to find on the
keyboard and mixing upper case for exported and snake case for separating
words might be mixing in too much.