docs: Update docs to use term target rather than assignable
Update docs to use term target rather than assignable, as assignable is at
times ambiguous: is the RHS value of an assignment assignable or the LHS? I
first tried using assign_target instead of target, but in some contexts, e.g.
type assertions it seems unfitting to call a value on the RHS assign_target.