frontend: Fix missing scrollbar in modal
Fix missing scrollbar in modal by adding a fixed height to the modal element.
This is needed for scrollbars to work. If 100% is used, it needs to also be
set on all parent element. <3 css. Adjust bottom padding in the modal to add
up correctly with the bottom padding of each example section.

Link: https://stackoverflow.com/a/32506818