frontend: Initialise editor content
Initialise editor content before JS is ready. This is the first part of a
series of commits trying to address slow and bad initial load experiences on
low bandwidth. It is best tested with network throtteling on and caching
disable in browser developper tooling.

The result of this update is that the sample evy source code loads before a
while before JS and evy.wasm are ready to go.