doctest: Exit with an error when errors occur
Ensure awk exits with an error when there is an error in an evy code
block. Prior to this, an error was printed but awk still exited with 0
allowing CI and `make doctest` to continue.

Emit a document line number when an error is reported to make it easier
to locate where the error is.