hermit: Add bash 5.2.15
Add bash 5.2.15 so we can make sure we are save when running on MacOS. An
initial version of the firebase-deploy script update I used associative
arrays which require a more modern bash version. This didn't make the final
cut, however, it seems prudent to ensure all scripts are run by the same
version of bash in this repo.