editor: Remove unused options and parameter types
Remove unused options and parameter types. This is an initial attempt at
better understanding the copied, open-source Yace editor which will likely
serve as base for an <evy-editor></evy-editor> web component. The exception
for a falsy selector constructor parameter has been removed as
document.querySelector already throws an exception or returns null which is
handled in the follow up exception. Less code better.