frontend: Clear URL fragment on update
Clear URL fragment on update on Evy source update, so that the URL fragment
stands for a specific unchanged source, and no hash stands for local edits.
In a follow-up commit, for an empty URL fragment we will try to load the
previous state saved in the browser's session store if present.