frontend: Save editor contents to session
Save editor contents to session storage so that accidental browser reloads to
not lose the contents. We attempt to reload the session if the URL fragment
is empty and the editor saves its content to session on every single
session.

Remove flicker on slow networks by rearranging editor initialization and
initial source code loading: Now the editor only gets initialized when the
source code is ready which reduces its empty state.

Remove scrollbar flicker in Evy editor on Firefox. In Firefox for
`overflow:auto` with overflowing content the scrollbar is briefly showed on
load then hidden. Unfortunately we update the contents of the editor twice
when found in session storage - once as part of index.html without syntax
highlighting, once with syntax highlighting when the editor and evy.wasm are
loaded and parsed (a good while later on slow networks without caching). To
remove the flicker we add a `.noscrollbar` class during the loading phase
that gets removed when the editor is ready.

Issue: https://github.com/evylang/evy/issues/227