scripts: Reorder functions for readability
Reorder functions for readability: `update_links` calls update
`update_links_in_file` and now comes first. There is no functional change.
This is more readable if you scan the file top to bottom. The reason it
wasn't done in the first place was because I wasn't sure if I had to declare
and export the function before I could use it. I wanted to get things to work
first and then make them pretty, but forgot about making things pretty in the
end - so let's do it now.