frontend: Refactor confetti module
Refactor confetti module for readability and finding my JS style. The JS used
in Evy is optimized for readability for Go developers so we'll take it easy
on the functional idiom - arrow functions, filter/map/reduce, etc.

There are also a couple of minor implementation fixes, such as improved side
offset for confetti, confetti size depending on viewport height, rather than
viewport with as we use in Evy. Options to configure confetti colors, print
and count.