deploy: Refactor deployment URLs in firebase-deploy
Refactor the generation of the site deployment URLs for the GitHub
environment to remove the chunks of repetitive JSON at the top of the
script and generate the URLs directly without JSON for the live
deployments.

The static chunk of data at the top of the script was a little
in-your-face and seemed rather repetitive, and in hindsight, probably
unnecessary when a simple bit of code could replace it.