Split the cicd workflow into stage and prod workflows
Split up the cicd workflow so that a PR does not show a bunch of
irrelevant skipped steps on it. The new `stage` workflow is triggered on
a pull request and can be called from another workflow. The `prod`
workflow is triggered on a push to `main`, calls the `stage` workflow
and does the deployment to prod.

Docs: https://docs.github.com/en/actions/using-workflows/reusing-workflows