🐒 github: Re-instate howl job (#247)
Re-instate howl job so we are notified on Slack and Discord when an evy
build on the `main` branch breaks. Instead of using the GitHub action we
are using the hermitised howl command so that it is more obvious on how to
run it locally.

Pull-Request: https://github.com/evylang/evy/pull/247