deploy: Rewrite site generation in Go
Rewrite the part of the `firebase-deploy` shell script that generates
the site in Go. To properly track the file renames in arbitrary
directories was going to get messy in shell, so this does it in Go.

The site generator does:
* Optionally rewrite URLs in `.html` files to some top-level paths to
  subdomains (discord, docs, learn, and play)
* Optionally rename `.js`, `.css` and `.wasm` files to include a
  cache-busting short hash of the content in the filename, and update
  any references to the renamed files in any `.html` files. These
  references include the stylesheet and javascript HTML references as
  well as an embedded javascript importmap and our custom embedded
  wasmImport map.

This adds a restriction that the site must not symlink directories -
instead it should create a directory and symlink the contents. Following
directory symlinks adds a lot of complexity for detecting loops,
recursively walking subtrees and tracking renames across that recursion
that it is much simpler to forbid directory symlinks.

The generated site can be used to deploy to a live firebase channel
(with subdomain rewriting and cache busting), to a PR channel (without
either) or to be embedded in a Go binary (all symlinks removed, no cache
busting or subdomain rewriting).