evy: Add serve command
Add `evy serve` command, starting a web server, serving the embedded evy website.

Because go embed does not work with symlinks we copy the frontend to out/embed
with the gen-site tool before embedding, which removes all symlinks. For this
to work with the current code base we first need built wasm with tinygo which
is slow. In order to avoid slow builds for common tasks like make go-test we
add the `embed-slim` and `install-slim` make targets that only embed a
placeholder page in the evy cli and therefor build a lot faster.

Co-authored-by: Cam Hutchison <camh@xdna.net>