🚚 build: Rename scripts directory to build-tools (#258)
The tools in the `scripts/` directory are for building and deploying
evy, but they wont always be scripts. In future, we'll have some Go
programs here instead of hacky scripts. Rename the directory to take
that into account, and update all references to the `scripts/` directory
to now point at `build-tools/`.

Pull-request: https://github.com/evylang/evy/pull/258