🚚 frontend: Move css/play.css to play/css/index.css (#263)
Move css/play.css to play/css/index.css. This was a workaround when we had the
bash script doing all firebase deployment and cache busting. Now that all of
this is addressed and fixed in the beautiful build-tools/site-gen go program
we can do this fancy stuff. Yay 🎉.

Pull-Request: https://github.com/evylang/evy/pull/263