frontend: Move some CSS rules around for reuse
Move some CSS rules around for reuse. Remove a couple of properties as well.
Now that we have screenshot testing I'm feeling ok to do this 🎉 I should
write a little script that tries removing every single property in all the
CSS files and lists the ones that keep the tests working.