tools: Update prettier config
Update prettier config to match defaults: Only update white-space according to
CSS rules. This means that only where we know with default CSS that adding
whitespace will make no difference will that's where we add it. Notably: keep
inline elements inlined. This popped up in a follow code change where
generate HTML documentation was formatted with each link on a single line
making it somewhat hard to read. That's not what we want. HTML should also be
made as readable as possilbe because "View Source" should be a viable
resource to learn about web development for evy.