frontend: Add generated HTML docs
We'll keep the generated HTML files in the repository because it will keep us
honest to keep creating readable HTML via reviews, we also have the full
static website there to be enjoyed and we also should notice unexpected
changes caused by the md build-tool.

We've extended the `check-uptodate` and clean targets to remove and rebuild
generated content and ensure the repo is still clean.