frontend: Fix ios rendering bug
Fix ios rendering bug where the height of an `<embed>` was different on
different browsers - it was higher on Safari and lower on FF and Chrome. The
result was a landing page with just a big evy-e and nothing else visible.
Thank you so much @starkcoffe for pointing this out.