frontend: Refactor header CSS
Refactor header CSS to be reuse-able in docs CSS. This is a preparatory step,
moving the header background into a CSS variable to be reused and maybe in
the end set by theme. Additionally we have reworked the meaning of .mobile
and .desktop classes as the previous setup caused issues with the default of
`display:inherit` overriding other display properties when set through other
classes.