frontend: Rework play header css
Refactor playground header CSS and move to shared location for re-use in docs
and other sites. We've pulled the style out into frontend/css/header.css and
added a minimal header.html file side-by-side for demonstration purposes.

We've also rearranged the header for mobile - the evy icon on the right links
to the homepage now and the left hand icon is an ordinary hamburger icon.

Snapshots of the playground had to all be updated. Pixels have been moved
around a little bit, neither is 100% exact to design but I believe the
current CSS structures is clearer and more correct using flexbox for
everything.