bytecode: Implement equality and inequality operators
The equality operators are the only operators that can
operate across all evy types. Therefore it makes sense
to utilise the value interface and its Equals function
to generically handle this expression for all evy types.
The alternative is to create separate equality and
inequality opcodes for each evy type, which would
introduce many new opcodes and repeated code in this
package.

Co-authored-by: joshcarp <joshcarp@users.noreply.github.com>
Co-authored-by: pgmitche <pgmitche@users.noreply.github.com>