frontend: Rename all sidemenu to sidebar
Rename all sidemenu to sidebar is this seems to be the more commonly used
term. On the GitHub wikis for example, _sidebar.md holds the contents of the
sidebar. In a follow up commit we will use the same convention for providing
the docs sidebar and for consistency we will start renaming here as a
preparatory step.