docs: Generate sidebar (generated)
Generate sidebar from _sidebar.md file: Each *.md file not prefixed with `_`
becomes a *.html file, for example spec.md becomes spec.html. For each of
these HTML pages the navigation section on the left side has the contents of
the `_sidebar.md` file from the **same** directory as the original .md file.

The contents of `_sidebar.md` are taken as they are _unless_ they are part of
a Markdown List where the first content of the list item is a relative link to
another markdown file. In that case the list item is expanded to the original link
followed by List of all level 2 and level 3 headings of that file nested within
each other. For example these `_sidebar.md` contents:

	# Documentation

	- [Language Specification](spec.md)
	- [Built-ins](builtins.md)

become

	# Documentation

	- [Language Specification](spec.md)
		- [Syntax Grammar](spec.md#syntax-grammar)
   			- [WSN Syntax Grammar](spec.md#wsn-syntax-grammar)
   			- [Evy Syntax Grammar](spec.md#evy-syntax-grammar)
		- [Comments](spec.md#comments)
		- [Types](spec.md#types)
		- ...
	- [Built-ins](builtins.md)
		- [Input and Output](builtins.md#input-and-output)
   			- [print](builtins.md#print),
   			- [read](builtins.md#read),
   			- [cls](builtins.md#cls),
   			- [printf](builtins.md#printf)
		- [Types](builtins.md#types)
     		- [len](builtins.md#len)
     		- [typeof](builtins.md#typeof)
     	- ...

Also add CSS styling in of expanded sidebar to this commit.
We will add logic to expand and collapse heading in follow-up commit.

Gen-command: make docs
Commit-group: 2/2