frontend: Update SVG files with version number (generated)
Update SVG files with version number in filename so we can include them in our
long expire cache policy. While at it, change the evy-e.svg DOM node from
<embed> (not officially supported, dates back to Flash days), to <object>.
This also explains previous strange behavior with embed on ios. We cannot
just use an <img> for this SVG file as that does not render it interactive.

Link: https://vecta.io/blog/best-way-to-embed-svg
Link: https://developer.mozilla.org/en-US/docs/Learn/HTML/Multimedia_and_embedding/Adding_vector_graphics_to_the_Web
Gen-command: make docs
Commit-group: 2/2