docsite: Center main content
Center main content rather than left-aligning it with different margin for
different breakpoints. For this to work, the main content needs to be wrapped
in a separate div - or at least I couldn't find a way to make centering with a
bit of minimum padding and scrollbars stuck to the right edge work.

Commit-group: 1/2