bytecode: Fix compile assignment
Add a test to cover assignment and fix a bug in the compiler. Which
was defining globals using the right hand side (value) of an
assignment, and compiling the left hand side (variable name).
This was not captured in the test suite because simple tests up
until this point only assigned the variable x to itself (to allow
the code compile).