🪚 build-tools: Rewrite the hrefs starting with / to apex (#308)
Rewrite the hrefs starting with / to apex. This only applies to "domain"
deployments to "evy.dev" or "evystage.dev" and not per PR preview
deployments. We are already re-writing the hrefs to /docs, /discord etc. to
https://docs.[DOMAIN], https://discord.[DOMAIN].

Now we are re-writing the hrefs starting with / to apex domains
(e.g. https://evy.dev/).

Pull-request: https://github.com/evylang/evy/pull/308