🧪 e2e: Attempt to fix flaky tests (#312)
Attempt to fix flaky tests, which is really hard as they are not reproducible
on either local run or PR deployments. The await for `networkidle` timed out
a couple of times on the docs tests, so I've attempted to just remove it. It
was added in the first place because the dialog loads an image
asynchronously, but maybe there is no need for an extra wait here.

We've also had a failure in the playground tests where that I saw in the same
way before I added the delay for the animation which should only take 200ms,
but maybe 300ms as enough extra little wiggle room.

Pull-request: https://github.com/evylang/evy/pull/312