docsite: Add syntax coloring
Add syntax coloring to docsite now that we have factored it out of the
playground. This requires that we replace ```evy:out and ```evy:err and
```evy:in tags with valid CSS classnames: `evy-out`, `evy-err` and `evy-in`
in build-tools/md.

Additionally we need to use the query selector `language-evy` to select all
evy code blocks, because rsc.io/markdown doesn't allow us to override the
classname used for codeblocks and reimplementing markdown.CodeBlock.PrintHTML
is not possible without copying big blocks of code - so we have to work with
the classname `language-evy` and `language-evy-out` for the time being.

We also need to link highlight.js and syntax.css with with the docs site and
update the the template in build-tools/md. docs/index.js has a little addition
of applying highlightEvy to all its code blocks.

Commit-group: 1/2