bytecode: Implement slice expressions
Support positive and negative slices in the compiler, introduce
OpNone to allow passing empty values for the start or end of a
slice to signal the vm to default the value. The vm will return
a new user error named ErrSlice if the start index of a slice
is greater than the end index.

Co-authored-by: joshcarp <joshcarp@users.noreply.github.com>
Co-authored-by: pgmitche <pgmitche@users.noreply.github.com>