bytecode: Implement if statements
Add two new opcodes (`OpJumpOnFalse`, `OpJump`) that give the vm the
ability to skip instructions. Add the `ChangeOperand` function to
`Instructions`, which allows the jump operand to be rewritten after
further instructions have been emitted to the bytecode (this process is
known as backfilling). The alternative to backfilling is a two-pass
compiler, which we discarded as too complex for now.

Co-authored-by: joshcarp <joshcarp@users.noreply.github.com>
Co-authored-by: pgmitche <pgmitche@users.noreply.github.com>