evy: Implement SVG support
Implement SVG support for the Evy CLI with flag `--svg-out`. In package
`cli/svg` we add the SVG runtime that implements a minimal subset of SVG to
output Evy drawings. Drawing, but no animations, are fully supported except
for startAngle and endAndle in ellipse. The SVG runtime is tested with the
sample drawing evy programs from the playground.

An SVG file generated by `evy run` consists of <rect>, <circle>, <line>,
<ellipse> and <text> elements. The elements are grouped within a <g> element
when they have the same style attributes, e.g. fill color, stroke color,
stroke width etc. When the style attributes are changed a new group is
started. If a group consists only of a single element, the group is omitted
and the styles are applied to the element itself, rather than the group. For example

```xml
<svg  viewBox="0 0 1000 1000" xmlns="http://www.w3.org/2000/svg" >
  <circle fill="red" cx="500" cy="500" r="100"></circle>
  <g fill="blue" >
    <rect x="200" y="700" width="100" height="300"></rect>
    <rect x="300" y="650" width="200" height="50"></rect>
  </g>
</svg>
```