evaluator: Add type to anyVal, stop using value.Type()
Add a `*parser.Type` field to `anyVal` to capture the dynamic type of
the `any` value, and stop using `value.Type()` to get that information.
`any` is the only place in evy that dynamic typing is used, so values
should not need to carry around any type information. When type
information is needed, it can come from the AST (which is only in two
places for error messages).

This allows a follow-up commit to remove the type information from
`arrayVal` and `mapVal` and to remove the `Type()` method from the
`value` interface, which starts the simplification of `value` with an
aim to make it part of a Go-ABI.