evaluator: Remove value.Type() method and types from arrayVal/mapVal
Remove the `Type()` method from `value` and remove the type fields from
`arrayVal` and `mapVal`. The previous commit reworked value types to
only need type information for `any` values and to otherwise get type
information from the AST if needed. This make the `Type()` method unused
as well as the type field of `arrayVal` and `mapVal`.