parser: Add an array repetition operator
Allow the `*` operator to be used on arrays with a number on the
right-hand side to repeat an array a number of times, producing an array
of the same type with repeated elements. This allows for creating arrays
of variable length without needing to write a loop.

e.g.

    n := 4
    a := [0] * n

will produce `[0 0 0 0]`.

There can be any number of elements in the array and they will all be
repeated. e.g. `[1 2] * 2` == `[1 2 1 2]`.

Update the spec to document this. Add an example to
`syntax_by_example.md`

Gen-command: make doc docs