evaluator: Implement array repetition
Implement construction of arrays via array repetition, a `BinaryExpr`
with an array value on the left and a number on the right. The result is
the left array repeated the right number of times. Zero times results in
an empty array. A non-integer value results in an error.

Add test cases for all distinct cases - basic, zero, precedence,
variable and error.