bytecode: Implement array repetition
Implement construction of arrays via array repetition, similar to the
evaluator. A new opcode, `OpArrayRepeat` is added for this operation
which the compiler emits based on the types of the arguments to the `*`
binary operator.

Add tests for some basic cases and the error case.

Disable some lint errors due to "complex" code. `vm.Run()` already
disabled one, but now there's another. The test case is not really even
complex. Stupid linters.