🐞 parser: Report error for quoted map keys (#339)
Report parse error for quoted map keys in map literals.
Before the following code did NOT report an error:

	m := {"a": true}
	print m

Now it does. The issue was that we accidentally printed `string` for string
literals in the parser which doubled up with the `string` keyword and made
AsIdent() work incorrectly for string literals which it now just passes
through.

Fixes: https://github.com/evylang/evy/issues/338
Pull-request: https://github.com/evylang/evy/pull/339