📖 docsite: Fix sidebar expansion for "cleanUrls" (#349)
Fix sidebar expansion for "cleanUrls", which is firebase config setting that
allows to load URLs without the ".html" extension. This looks neat in the
address bar but unfortunately broke the sidebar expansion on site reload or
even navigation within the page the a different anchor/hash. With this fix we
have sidebar expand correctly with and without the cleanURL setting.

Pull-request: https://github.com/evylang/evy/pull/349