firebase: Remove cleanUruls again
Remove cleanUruls again, firstly it causes a 301 redirect when trying to
access the .html pages which is where all the links point which is also
useful for local or basic fileserve deployment. Secondly it breaks anchor
navigation within the site and causes a brief flicker of the sidebar is the
redirected site is loaded from the firebase server again:
when loading /spec an anchor link to /spec.html#map causes a full page reload.

Lastly, Evy is about learning and understanding all things code. Making use of
some firebase setting to avoid typing ".html" makes things less clear.