learn: Fix HTML output for checkbox and radio
Fix HTML output for checkbox and radio by using "value" instead of "id". This
is the correct attribute that would get automatically submitted on on a form
submit, and it will also not clash with other questions should they be
present on the same page.

Secondly use "checkbox" instead of "radio" type input for multiple-choice
questions.