✨ cli: Add hidden "compile" command to CLI (#365)
Add a `compile` command to the CLI that is hidden, like the `tokenize`
and `parse` commands. `compile` outputs the bytecode for the evy program
specified on the command line (or stdin). It is intended for debugging
the compiler and for deeper understanding of what the bytecode for a
program looks like.

Pull-request: https://github.com/evylang/evy/pull/365