playsite: Add `editor=none` options to hash
Add `editor=none` options to hash in preparation for classroom work: I'd like
to show students a program in execution without giving away how it's
programmed - find out what this thing does by just observing it.

This is a hack. Truth is the playground is ripe for a re-think and re-write,
ideally based around web-components. But that's a big job and this will do
for now. I hope.