learn: Refactor Answerkey
Remove the requirement for question markdown files to be within a `questions`
subdirectory. Question markdown files will live side-by-side with exercise
markdown files.

Refactor Answerkey internals to make it easier to add new answers to an
existing answerkey. This will be used in a follow-up commit where we add the
answer of all the questions of a single exercise to the answerkey.