learn: Add CSS rules to HTML export
Add CSS rule for not-first-child <form> elements to HTML export of question.
This isn't used directly in the question, but will be used in the exercise.
Add a little bit of bottom-padding to the the form element.

Add some headings to the questions. All of this is prep work for having
multiple questions within one exercise.

Sneak in a some basic table styling currently only used in exercise
composition, but theoretically also useful if tables are used as part of
question.