all: Fix URLs in Markdown files
It is time again for a tools upgrade. I haven't run `make test-urls` in a
while, a target that ensures all external URLs are still available
periodically.

I now did and it brought to light that an inline link in the learn docs
doesn't follow the convention of external links in footnote rather than
inline style only. Another link in the example directory was outright
wrong. All fixed up now.