bin: Upgrade hermitised tools
Upgrade hermitised tools with

	hermit upgrade

Double-check that there is no new major release of firebase, node or openjre.
We're all good. As usual, golangci-lint gives us upgrade challenges to solve.
This time some disabled linters are deprecated "enough" to cause a warning
when listed as disabled, so we removed the following from the .golangci.yaml
config:

- deadcode
- exhaustivestruct
- golint
- ifshort
- interfacer
- maligned
- nosnakecase
- scopelint
- structcheck
- varcheck