💥 learn: Add directory exporting (#375)
Add directory exporting to the `learn` package and use it in the `levy` CLI in
order to process _all_ files. We add this to include all media (e.g. images)
when creating the target HTML structure but also to prepare for modelling
exercises, units and courses.

Pull-request: https://github.com/evylang/evy/pull/375