learn: Refactor learn.AnswerPath
Refactor AnswerPath in learn package to be be used as immutable type, without
pointer and to be Exported as it will be relevant to get the AnswerPath from
outside the learn package. During some refactors this refactor export tests
started failing because of MacOS .DS_Store file. Let's ignore it in the
golden tests.