learn: Add course catalog exporting
Add course catalog exporting as single JSON file. Catalog contains all
information to generate a question set for an exercise, quiz or unittest
identified by composite IDs. It also holds the maximumScore achievable for
each course and unit. An exercise has a predefined score of 10. This
will allow for progress tracking as percentage.