learn: Add Filename() method to model
Add Filename() method to model interface. This allows us to use
model.Filename() to create the HTML output files rather than relying on the
original `mdFiles` slice from which we created the models.

In a follow-up commit we break 1-1 relationship between .md and .html files
when we create sub-question from txtar files. For these types of questions we
require that the QuestionModel provides its (target) filename.