learn: Generate SubQuestions from question.md template
Generate SubQuestions from question.md template, where questions and
answers are all represented in .txtar files. We then "generate" a
sub-question from each possible correct answer.

This is a pretty messy extension to the existing codebase unfortunately, but
if we see clearer how the levy tool is used it might get easier to refactor
only the used features. Sorry.